refactor: move createReportViaApi to reportsPages/reportCreation.js
Follows the same pattern as createDashboardViaApi in dashCreation.js —
API creation helpers live alongside their page objects, not in apiCleanup.
