+++ tests/ui-testing/pages/apiCleanup.js
@@ -349,95 +349,6 @@ class APICleanup {
         }
     }
 
-    /**
-     * Create a report via API (v2) in the default folder.
-     * Fetches the first available dashboard and uses it as the report source.
-     * @param {string} reportName - Unique name for the report
-     * @param {string} folderId - Folder ID to create the report in (defaults to "default")
-     * @returns {Promise<Object>} Result with { success, reportName, error }
-     */
-    async createReportViaApi(reportName, folderId = 'default') {
-        testLogger.info('Creating report via API', { reportName, folderId });
-
-        try {
-            // Step 1: Fetch dashboards from the folder to get a dashboard ID
-            const dashboards = await this.fetchDashboardsInFolder(folderId);
-            if (dashboards.length === 0) {
-                testLogger.error('No dashboards found in folder, cannot create report', { folderId });
-                return { success: false, error: 'No dashboards available' };
-            }
-
-            const dashboard = dashboards[0];
-            // Unwrap MetaDashboard if data is nested under version key (e.g., "v8")
-            const inner = dashboard[`v${dashboard.version}`] || dashboard;
-            if (!inner.dashboardId && !inner.dashboard_id) {
-                testLogger.error('Dashboard missing required fields', { dashboard });
-                return { success: false, error: 'Dashboard missing dashboard_id' };
-            }
-
-            const dashboardId = inner.dashboardId || inner.dashboard_id;
-            const dashboardTabs = inner.tabs || [{ tabId: 'default' }];
-            testLogger.info('Using dashboard for report', { dashboardId, title: inner.title });
-
-            // Step 2: Build the report payload
-            const payload = {
-                name: reportName,
-                description: '',
-                dashboards: [{
-                    folder: folderId,
-                    dashboard: dashboardId,
-                    tabs: [dashboardTabs[0]?.tabId || dashboardTabs[0]?.tab_id || 'default'],
-                    variables: [],
-                    timerange: {
-                        type: 'relative',
-                        period: '30m',
-                        from: 0,
-                        to: 0
-                    },
-                    report_type: 'pdf',
-                    email_attachment_type: 'standard'
-                }],
-                destinations: [{ email: this.email }],
-                enabled: true,
-                imagePreview: false,
-                title: `Test Report ${reportName}`,
-                message: '',
-                orgId: this.org,
-                frequency: { interval: 1, type: 'once', cron: '' },
-                timezone: 'UTC',
-                timezoneOffset: 0,
-                lastTriggeredAt: null,
-                owner: this.email,
-                lastEditedBy: this.email
-            };
-
-            // Step 3: POST to create the report
-            const response = await this._fetch(
-                `${this.baseUrl}/api/v2/${this.org}/reports?folder=${encodeURIComponent(folderId)}`,
-                {
-                    method: 'POST',
-                    headers: {
-                        'Authorization': this.authHeader,
-                        'Content-Type': 'application/json'
-                    },
-                    body: JSON.stringify(payload)
-                }
-            );
-
-            if (!response.ok) {
-                const errorBody = await response.text();
-                testLogger.error('Failed to create report via API', { reportName, status: response.status, body: errorBody });
-                return { success: false, error: `HTTP ${response.status}: ${errorBody}` };
-            }
-
-            const result = await response.json();
-            testLogger.info('Report created via API', { reportName, result });
-            return { success: true, reportName };
-        } catch (error) {
-            testLogger.error('Failed to create report via API', { reportName, error: error.message });
-            return { success: false, error: error.message };
-        }
-    }
 
     /**
      * Fetch all pipelines
